NeuroMap/NeuroMap/requirements.txt
flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
google-genai==0.2.2